refactor: re-organize directories and update cli before merge
