--- src/_get-introspection.ts
@@ -1,6 +1,0 @@
-import { fetchIntrospection, saveJsonFormatted } from "./lib.ts";
-
-
-const introspection = await fetchIntrospection("https://rickandmortyapi.com/graphql");
-// const {queryType, mutationType} = getQueryAndMutationTypes(introspection);
-saveJsonFormatted(introspection, "introspection.json");
