+++ src/index.ts
@@ -1,4 +1,5 @@
 import { outrospectionToQueries } from "./converters.ts";
+import { queryCollectionToPostmanCollection } from "./format.ts";
 import { fetchIntrospection, saveJsonFormatted } from "./lib.ts";
 import { outrospect, outrospectionToJSON } from "./outrospector.ts";
 
@@ -11,4 +12,10 @@ export async function createPostmanCollection(
   const introspection = await fetchIntrospection(url, authorization);
   const outrospection = outrospect(introspection);
   const queryCollection = outrospectionToQueries(outrospection);
+  const postmanCollection = queryCollectionToPostmanCollection(
+    queryCollection,
+    url,
+    authorization,
+  );
+  return postmanCollection;
 }
